Modularize capture lifecycle orchestration
Separate meeting disclosure, capture finalization, and resume recovery from the recording start hook while preserving existing exports and behavior.
