+++ apps/desktop/src/stt/meeting-disclosure.ts
@@ -1,0 +1,192 @@
+import { commands as detectCommands } from "@anlg/plugin-detect";
+import { sonnerToast } from "@anlg/ui/components/ui/toast";
+
+export const MEETING_DISCLOSURE_MESSAGE =
+  "I'm using Anarlog to record and transcribe this meeting. https://anarlog.so";
+
+const MEETING_DISCLOSURE_MAX_ATTEMPTS = 30;
+const MEETING_DISCLOSURE_RETRY_INTERVAL_MS = 1_000;
+const SLACK_BUNDLE_IDS = new Set([
+  "com.slack.Slack",
+  "com.tinyspeck.slackmacgap",
+]);
+
+type MeetingDisclosureOutcome =
+  | { status: "sent" }
+  | { status: "notSent"; reason: string }
+  | { status: "cancelled" };
+
+type MeetingDisclosureAttemptOutcome =
+  | { status: "sent" }
+  | { status: "notSent"; reason: unknown }
+  | { status: "cancelled" };
+
+type MeetingDisclosureTask = {
+  cancelled: boolean;
+  restartWhenSettled?: () => boolean;
+  status: "sending" | "sent";
+};
+
+const meetingDisclosureTasks = new Map<string, MeetingDisclosureTask>();
+
+function meetingDisclosureFailure(reason: unknown): MeetingDisclosureOutcome {
+  const detail = reason instanceof Error ? reason.message : String(reason);
+  console.warn("[listener] meeting disclosure was not sent", reason);
+  sonnerToast.warning(
+    "Recording started, but Anarlog could not post the meeting chat disclosure.",
+    { id: "meeting-disclosure-send-failed" },
+  );
+  return { status: "notSent", reason: detail };
+}
+
+async function attemptMeetingRecordingDisclosure(
+  isCancelled: () => boolean,
+): Promise<MeetingDisclosureAttemptOutcome> {
+  if (isCancelled()) {
+    return { status: "cancelled" };
+  }
+
+  let micAppsResult: Awaited<
+    ReturnType<typeof detectCommands.listMicUsingApplications>
+  >;
+
+  try {
+    micAppsResult = await detectCommands.listMicUsingApplications();
+  } catch (error) {
+    return isCancelled()
+      ? { status: "cancelled" }
+      : { status: "notSent", reason: error };
+  }
+
+  if (isCancelled()) {
+    return { status: "cancelled" };
+  }
+
+  if (micAppsResult.status === "error") {
+    return { status: "notSent", reason: micAppsResult.error };
+  }
+
+  const micActiveBundleIds = [
+    ...new Set(micAppsResult.data.map((app) => app.id.trim()).filter(Boolean)),
+  ];
+  if (!micActiveBundleIds.some((bundleId) => SLACK_BUNDLE_IDS.has(bundleId))) {
+    return {
+      status: "notSent",
+      reason: "no mic-active Slack app was found",
+    };
+  }
+
+  if (isCancelled()) {
+    return { status: "cancelled" };
+  }
+
+  let result: Awaited<ReturnType<typeof detectCommands.sendMeetingChatMessage>>;
+
+  try {
+    result = await detectCommands.sendMeetingChatMessage(
+      MEETING_DISCLOSURE_MESSAGE,
+      micActiveBundleIds,
+    );
+  } catch (error) {
+    return isCancelled()
+      ? { status: "cancelled" }
+      : { status: "notSent", reason: error };
+  }
+
+  if (result.status === "error") {
+    return isCancelled()
+      ? { status: "cancelled" }
+      : { status: "notSent", reason: result.error };
+  }
+
+  if (result.data.sent) {
+    return { status: "sent" };
+  }
+
+  if (isCancelled()) {
+    return { status: "cancelled" };
+  }
+
+  return {
+    status: "notSent",
+    reason:
+      result.data.warnings.join("; ") || "meeting chat mutation was rejected",
+  };
+}
+
+export async function sendMeetingRecordingDisclosure({
+  isCancelled = () => false,
+  maxAttempts = MEETING_DISCLOSURE_MAX_ATTEMPTS,
+  retryIntervalMs = MEETING_DISCLOSURE_RETRY_INTERVAL_MS,
+}: {
+  isCancelled?: () => boolean;
+  maxAttempts?: number;
+  retryIntervalMs?: number;
+} = {}): Promise<MeetingDisclosureOutcome> {
+  let lastFailureReason: unknown = "meeting chat disclosure was not sent";
+
+  for (let attempt = 0; attempt < Math.max(1, maxAttempts); attempt += 1) {
+    const outcome = await attemptMeetingRecordingDisclosure(isCancelled);
+    if (outcome.status !== "notSent") {
+      return outcome;
+    }
+
+    lastFailureReason = outcome.reason;
+    if (attempt + 1 < Math.max(1, maxAttempts)) {
+      await new Promise<void>((resolve) => {
+        setTimeout(resolve, retryIntervalMs);
+      });
+      if (isCancelled()) {
+        return { status: "cancelled" };
+      }
+    }
+  }
+
+  return meetingDisclosureFailure(lastFailureReason);
+}
+
+export function startMeetingRecordingDisclosure(
+  sessionId: string,
+  isListening: () => boolean,
+) {
+  const existingTask = meetingDisclosureTasks.get(sessionId);
+  if (existingTask) {
+    if (existingTask.status === "sending" && existingTask.cancelled) {
+      existingTask.restartWhenSettled = isListening;
+    }
+    return;
+  }
+
+  const task: MeetingDisclosureTask = {
+    cancelled: false,
+    status: "sending",
+  };
+  meetingDisclosureTasks.set(sessionId, task);
+
+  void sendMeetingRecordingDisclosure({
+    isCancelled: () => task.cancelled || !isListening(),
+  }).then((outcome) => {
+    if (meetingDisclosureTasks.get(sessionId) !== task) {
+      return;
+    }
+
+    if (outcome.status === "sent") {
+      task.status = "sent";
+    } else {
+      const restartWhenSettled = task.restartWhenSettled;
+      meetingDisclosureTasks.delete(sessionId);
+      if (restartWhenSettled?.()) {
+        startMeetingRecordingDisclosure(sessionId, restartWhenSettled);
+      }
+    }
+  });
+}
+
+export function cancelMeetingRecordingDisclosure(sessionId: string) {
+  const task = meetingDisclosureTasks.get(sessionId);
+  if (!task || task.status === "sent") {
+    return;
+  }
+
+  task.cancelled = true;
+}
